Feat: created User Entity
